refactor katkem into lib
